Actualizacion de repositorios

+++ src/Iconos.Geograficos.Api/Iconos.Geograficos.Api/Model/IconosGeograficosViewModel.cs
@@ -10,7 +10,7 @@
        
         public DateTime FechaCreacion { get; set; }
 
-        public int Altura { get; set; }
+        public decimal Altura { get; set; }
 
         public string Historia { get; set; }
